refactor(client): Transition 컴포넌트 util폴더로 이동
